tempban core changes to .gitignore, TGNSCommon and new config and plugin files
@lancehilliard I'm stuck on what all needs to be in place for the
sv_tempban function to be available in the console for tempadmins, and
I'm not positive which file(s) I need to change.

Since I can't seem to have it available in the console, I can't tell
whether I'm using the GetPlayerMatching correctly. My main goal is to
get a message to be displayed when the correct roles use sv_tempban,
then I will work on the ban part.